Split tests/integration/bigquery.spec.ts
