+++ tests/integration/bigquery.spec.ts
@@ -1,222 +1,15 @@
 import { expect } from "chai";
-import * as fs from "fs-extra";
-import Long from "long";
-import * as path from "path";
 
 import * as dfapi from "df/cli/api";
-import * as dbadapters from "df/cli/api/dbadapters";
 import { BigQueryDbAdapter } from "df/cli/api/dbadapters/bigquery";
 import { ExecutionSql } from "df/cli/api/dbadapters/execution_sql";
-import { INTEGRATION_TEST_LOCATION, INTEGRATION_TEST_PROJECT } from "df/cli/index_test_base";
-import { targetAsReadableString } from "df/core/targets";
 import { dataform } from "df/protos/ts";
 import { suite, test } from "df/testing";
-import { compile, dropAllTables, getTableRows, keyBy } from "df/tests/integration/utils";
+import { cleanWarehouse, compile } from "df/tests/integration/utils";
 
-suite("@dataform/integration/bigquery", { parallel: true }, ({ before, after }) => {
+suite("@dataform/integration/bigquery", { parallel: true }, () => {
   const credentials = dfapi.credentials.read("test_credentials/bigquery.json");
-  let dbadapter: BigQueryDbAdapter;
-
-  before("create adapter", async () => {
-    dbadapter = new BigQueryDbAdapter(credentials);
-  });
-
-  suite("run", { parallel: true }, () => {
-    test("project e2e", { timeout: 60000 }, async () => {
-      const compiledGraph = await compile("tests/integration/bigquery_project", "project_e2e");
-
-      // Drop all the tables before we do anything.
-      await cleanWarehouse(compiledGraph, dbadapter);
-
-      // Drop schemas to make sure schema creation works.
-      await dbadapter.execute(
-        `drop schema if exists \`${INTEGRATION_TEST_PROJECT}.df_integration_test_project_e2e\` cascade`
-      );
-
-      // Run the project.
-      const executionGraph = await dfapi.build(compiledGraph, {}, dbadapter);
-      const executedGraph = await dfapi.run(dbadapter, executionGraph).result();
-
-      const actionMap = keyBy(executedGraph.actions, v => targetAsReadableString(v.target));
-      expect(Object.keys(actionMap).length).eql(20);
-
-      // Check the status of action execution.
-      const expectedFailedActions = [
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_assertions_project_e2e.example_assertion_fail`,
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_project_e2e.example_operation_partial_fail`
-      ];
-      for (const actionName of Object.keys(actionMap)) {
-        const expectedResult = expectedFailedActions.includes(actionName)
-          ? dataform.ActionResult.ExecutionStatus.FAILED
-          : dataform.ActionResult.ExecutionStatus.SUCCESSFUL;
-        expect(actionMap[actionName].status).equals(
-          expectedResult,
-          JSON.stringify(actionMap[actionName], null, 4)
-        );
-      }
-
-      expect(
-        actionMap[
-          `${INTEGRATION_TEST_PROJECT}.df_integration_test_assertions_project_e2e.example_assertion_fail`
-        ].tasks[1].errorMessage
-      ).to.eql("bigquery error: Assertion failed: query returned 1 row(s).");
-
-      expect(
-        actionMap[
-          `${INTEGRATION_TEST_PROJECT}.df_integration_test_project_e2e.example_operation_partial_fail`
-        ].tasks[0].errorMessage
-      ).to.eql("bigquery error: Query error: Unrecognized name: invalid_column at [3:8]");
-    });
-
-    test("incremental tables", { timeout: 60000 }, async () => {
-      const compiledGraph = await compile(
-        "tests/integration/bigquery_project",
-        "incremental_tables"
-      );
-
-      // Drop all the tables before we do anything.
-      await cleanWarehouse(compiledGraph, dbadapter);
-
-      // Run two iterations of the project.
-      const adapter = new ExecutionSql(
-        compiledGraph.projectConfig,
-        compiledGraph.dataformCoreVersion
-      );
-      for (const runIteration of [
-        {
-          runConfig: {
-            actions: ["example_incremental", "example_incremental_merge"],
-            includeDependencies: true
-          },
-          expectedIncrementalRows: 3,
-          expectedIncrementalMergeRows: 2
-        },
-        {
-          runConfig: {
-            actions: ["example_incremental", "example_incremental_merge"]
-          },
-          expectedIncrementalRows: 5,
-          expectedIncrementalMergeRows: 2
-        }
-      ]) {
-        const executionGraph = await dfapi.build(compiledGraph, runIteration.runConfig, dbadapter);
-        const runResult = await dfapi.run(dbadapter, executionGraph).result();
-        expect(dataform.RunResult.ExecutionStatus[runResult.status]).eql(
-          dataform.RunResult.ExecutionStatus[dataform.RunResult.ExecutionStatus.SUCCESSFUL]
-        );
-        const [incrementalRows, incrementalMergeRows] = await Promise.all([
-          getTableRows(
-            {
-              database: INTEGRATION_TEST_PROJECT,
-              schema: "df_integration_test_incremental_tables",
-              name: "example_incremental"
-            },
-            adapter,
-            dbadapter
-          ),
-          getTableRows(
-            {
-              database: INTEGRATION_TEST_PROJECT,
-              schema: "df_integration_test_incremental_tables",
-              name: "example_incremental_merge"
-            },
-            adapter,
-            dbadapter
-          )
-        ]);
-        expect(incrementalRows.length).equals(runIteration.expectedIncrementalRows);
-        expect(incrementalMergeRows.length).equals(runIteration.expectedIncrementalMergeRows);
-      }
-    });
-
-    test("dataset metadata set correctly", { timeout: 60000 }, async () => {
-      const compiledGraph = await compile("tests/integration/bigquery_project", "dataset_metadata");
-
-      // Drop all the tables before we do anything.
-      await cleanWarehouse(compiledGraph, dbadapter);
-
-      // Run the project.
-      const executionGraph = await dfapi.build(
-        compiledGraph,
-        {
-          actions: ["example_incremental", "example_view"],
-          includeDependencies: true
-        },
-        dbadapter
-      );
-      const runResult = await dfapi.run(dbadapter, executionGraph).result();
-      expect(dataform.RunResult.ExecutionStatus[runResult.status]).eql(
-        dataform.RunResult.ExecutionStatus[dataform.RunResult.ExecutionStatus.SUCCESSFUL]
-      );
-
-      // Check expected metadata.
-      for (const expectedMetadata of [
-        {
-          target: {
-            database: INTEGRATION_TEST_PROJECT,
-            schema: "df_integration_test_dataset_metadata",
-            name: "example_incremental"
-          },
-          expectedDescription: "An incremental table",
-          expectedFields: [
-            dataform.Field.create({
-              description: "the timestamp",
-              name: "user_timestamp",
-              primitive: dataform.Field.Primitive.INTEGER
-            }),
-            dataform.Field.create({
-              description: "the id",
-              name: "user_id",
-              primitive: dataform.Field.Primitive.INTEGER
-            }),
-            dataform.Field.create({
-              name: "nested_data",
-              description: "some nested data with duplicate fields",
-              struct: dataform.Fields.create({
-                fields: [
-                  dataform.Field.create({
-                    description: "nested timestamp",
-                    name: "user_timestamp",
-                    primitive: dataform.Field.Primitive.INTEGER
-                  }),
-                  dataform.Field.create({
-                    description: "nested id",
-                    name: "user_id",
-                    primitive: dataform.Field.Primitive.INTEGER
-                  })
-                ]
-              })
-            })
-          ],
-          expectedLabels: {}
-        },
-        {
-          target: {
-            database: INTEGRATION_TEST_PROJECT,
-            schema: "df_integration_test_dataset_metadata",
-            name: "example_view"
-          },
-          expectedDescription: "An example view",
-          expectedFields: [
-            dataform.Field.create({
-              description: "val doc",
-              name: "val",
-              primitive: dataform.Field.Primitive.INTEGER
-            })
-          ],
-          expectedLabels: {
-            label1: "val1",
-            label2: "val2"
-          }
-        }
-      ]) {
-        const metadata = await dbadapter.table(expectedMetadata.target);
-        expect(metadata.description).to.equal(expectedMetadata.expectedDescription);
-        expect(metadata.fields).to.deep.equal(expectedMetadata.expectedFields);
-        expect(metadata.labels).to.deep.equal(expectedMetadata.expectedLabels);
-      }
-    });
-  });
+  const dbadapter = new BigQueryDbAdapter(credentials);
 
   test("run unit tests", async () => {
     const compiledGraph = await compile("tests/integration/bigquery_project", "unit_tests");
@@ -251,125 +44,6 @@ suite("@dataform/integration/bigquery", { parallel: true }, ({ before, after }) 
       },
       { name: "test a view", successful: true }
     ]);
-  });
-
-  suite("evaluate", async () => {
-    test("evaluate from valid compiled graph as valid", async () => {
-      // Create and run the project.
-      const compiledGraph = await compile("tests/integration/bigquery_project", "evaluate");
-      const executionGraph = await dfapi.build(compiledGraph, {}, dbadapter);
-      await dfapi.run(dbadapter, executionGraph).result();
-
-      const view = keyBy(compiledGraph.tables, t => targetAsReadableString(t.target))[
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_evaluate.example_view`
-      ];
-      let evaluations = await dbadapter.evaluate(dataform.Table.create(view));
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-
-      const materializedView = keyBy(compiledGraph.tables, t => targetAsReadableString(t.target))[
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_evaluate.example_materialized_view`
-      ];
-      evaluations = await dbadapter.evaluate(dataform.Table.create(materializedView));
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-
-      const table = keyBy(compiledGraph.tables, t => targetAsReadableString(t.target))[
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_evaluate.example_table`
-      ];
-      evaluations = await dbadapter.evaluate(dataform.Table.create(table));
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-
-      const operation = keyBy(compiledGraph.operations, t => targetAsReadableString(t.target))[
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_evaluate.example_operation`
-      ];
-      evaluations = await dbadapter.evaluate(dataform.Operation.create(operation));
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-
-      const assertion = keyBy(compiledGraph.assertions, t => targetAsReadableString(t.target))[
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_assertions_evaluate.example_assertion_pass`
-      ];
-      evaluations = await dbadapter.evaluate(dataform.Assertion.create(assertion));
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-
-      const incremental = keyBy(compiledGraph.tables, t => targetAsReadableString(t.target))[
-        `${INTEGRATION_TEST_PROJECT}.df_integration_test_evaluate.example_incremental`
-      ];
-      evaluations = await dbadapter.evaluate(dataform.Table.create(incremental));
-      expect(evaluations.length).to.equal(2);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-      expect(evaluations[1].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-    });
-
-    test("variable persistence validated correctly", async () => {
-      const target = (name: string) => ({
-        schema: "df_integration_test",
-        name,
-        database: INTEGRATION_TEST_PROJECT
-      });
-
-      let evaluations = await dbadapter.evaluate(
-        dataform.Table.create({
-          enumType: dataform.TableType.TABLE,
-          preOps: ["declare var string; set var = 'val';"],
-          query: "select var as col;",
-          target: target("example_valid_variable")
-        })
-      );
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.SUCCESS
-      );
-
-      evaluations = await dbadapter.evaluate(
-        dataform.Table.create({
-          enumType: dataform.TableType.TABLE,
-          query: "select var as col;",
-          target: target("example_invalid_variable")
-        })
-      );
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.FAILURE
-      );
-    });
-
-    test("invalid table fails validation and error parsed correctly", async () => {
-      const evaluations = await dbadapter.evaluate(
-        dataform.Table.create({
-          enumType: dataform.TableType.TABLE,
-          query: "selects\n1 as x",
-          target: {
-            name: "EXAMPLE_ILLEGAL_TABLE",
-            database: "df_integration_test"
-          }
-        })
-      );
-      expect(evaluations.length).to.equal(1);
-      expect(evaluations[0].status).to.equal(
-        dataform.QueryEvaluation.QueryEvaluationStatus.FAILURE
-      );
-      expect(
-        dataform.QueryEvaluationError.ErrorLocation.create(evaluations[0].error.errorLocation)
-      ).eql(dataform.QueryEvaluationError.ErrorLocation.create({ line: 1, column: 1 }));
-    });
   });
 
   suite("publish tasks", { parallel: true }, async () => {
@@ -409,60 +83,6 @@ suite("@dataform/integration/bigquery", { parallel: true }, ({ before, after }) 
     });
   });
 
-  suite("execute", { parallel: true }, async () => {
-    test("returned metadata includes jobReference and statistics", async () => {
-      const query = `select 1 as test`;
-      const { metadata } = await dbadapter.execute(query);
-      const { bigquery: bqMetadata } = metadata;
-      expect(bqMetadata).to.have.property("jobId");
-      expect(bqMetadata.jobId).to.match(
-        /^dataform-[0-9A-Fa-f]{8}(?:-[0-9A-Fa-f]{4}){3}-[0-9A-Fa-f]{12}$/
-      );
-      expect(bqMetadata).to.have.property("totalBytesBilled");
-      expect(bqMetadata.totalBytesBilled).to.eql(Long.fromNumber(0));
-      expect(bqMetadata).to.have.property("totalBytesProcessed");
-      expect(bqMetadata.totalBytesProcessed).to.eql(Long.fromNumber(0));
-    });
-
-    test("configured job prefix is added to table names", async () => {
-      const query = `select 1 as test`;
-      const { metadata } = await dbadapter.execute(query, { bigquery: { jobPrefix: "jobPrefix" } });
-      const { bigquery: bqMetadata } = metadata;
-      expect(bqMetadata).to.have.property("jobId");
-      expect(bqMetadata.jobId).to.match(
-        /^dataform-jobPrefix-[0-9A-Fa-f]{8}(?:-[0-9A-Fa-f]{4}){3}-[0-9A-Fa-f]{12}$/
-      );
-    });
-
-    suite("query limits work", { parallel: true }, async () => {
-      const query = `
-        select 1 union all
-        select 2 union all
-        select 3 union all
-        select 4 union all
-        select 5`;
-
-      for (const options of [
-        { interactive: true, rowLimit: 2 },
-        { interactive: false, rowLimit: 2 },
-        { interactive: true, byteLimit: 30 },
-        { interactive: false, byteLimit: 30 }
-      ]) {
-        test(`with options=${JSON.stringify(options)}`, async () => {
-          const { rows } = await dbadapter.execute(query, options);
-          expect(rows).to.eql([
-            {
-              f0_: 1
-            },
-            {
-              f0_: 2
-            }
-          ]);
-        });
-      }
-    });
-  });
-
   test("search", async () => {
     const compiledGraph = await compile("tests/integration/bigquery_project", "search");
 
@@ -493,138 +113,4 @@ suite("@dataform/integration/bigquery", { parallel: true }, ({ before, after }) 
     expect(partialSearch.length).equals(2);
     expect(columnSearch.length).greaterThan(0);
   });
-
-  test("JiT execution e2e", { timeout: 120000 }, async () => {
-    // Create a simple project with a JiT table
-    const projectDir = "tests/integration/jit_project";
-    if (fs.existsSync(projectDir)) {
-      fs.removeSync(projectDir);
-    }
-    fs.mkdirpSync(path.join(projectDir, "definitions"));
-    fs.writeFileSync(
-      path.join(projectDir, "workflow_settings.yaml"),
-      `
-defaultProject: ${INTEGRATION_TEST_PROJECT}
-defaultLocation: ${INTEGRATION_TEST_LOCATION}
-defaultDataset: df_integration_test_jit
-`
-    );
-    fs.writeFileSync(
-      path.join(projectDir, "definitions/jit_table.js"),
-      `publish("jit_table", { type: "table" }).jitCode(async (jctx) => "SELECT 1 as id")`
-    );
-
-    // Mock @dataform/core to avoid npm install and 403 error
-    const nodeModulesDir = path.join(projectDir, "node_modules", "@dataform", "core");
-    fs.mkdirpSync(nodeModulesDir);
-    const coreBundlePath = path.resolve("packages/@dataform/core/bundle.js");
-    fs.copyFileSync(coreBundlePath, path.join(nodeModulesDir, "bundle.js"));
-    const corePackageJsonPath = path.resolve("packages/@dataform/core/package.json");
-    fs.copyFileSync(corePackageJsonPath, path.join(nodeModulesDir, "package.json"));
-    // We also need a package.json in the project root to bypass the dataformCoreVersion check
-    fs.writeFileSync(
-      path.join(projectDir, "package.json"),
-      JSON.stringify({ dependencies: { "@dataform/core": "3.0.0-alpha.0" } })
-    );
-
-    try {
-      const compiledGraph = await dfapi.compile({ projectDir });
-      
-      // Drop dataset to start fresh
-      await dbadapter.execute(
-        `drop schema if exists \`${INTEGRATION_TEST_PROJECT}.df_integration_test_jit\` cascade`
-      );
-
-      const executionGraph = await dfapi.build(compiledGraph, {}, dbadapter);
-      const runResult = await dfapi.run(dbadapter, executionGraph, { projectDir }).result();
-
-      expect(dataform.RunResult.ExecutionStatus[runResult.status]).eql(
-        dataform.RunResult.ExecutionStatus[dataform.RunResult.ExecutionStatus.SUCCESSFUL]
-      );
-
-      const rows = await dbadapter.execute(`SELECT * FROM \`${INTEGRATION_TEST_PROJECT}.df_integration_test_jit.jit_table\``).then(res => res.rows);
-      expect(rows).to.eql([{ id: 1 }]);
-    } finally {
-      if (fs.existsSync(projectDir)) {
-        fs.removeSync(projectDir);
-      }
-    }
-  });
-
-  test("JiT dry run integration", { timeout: 120000 }, async () => {
-    // Create a simple project with a JiT table
-    const projectDir = "tests/integration/jit_dry_run_project";
-    if (fs.existsSync(projectDir)) {
-      fs.removeSync(projectDir);
-    }
-    fs.mkdirpSync(path.join(projectDir, "definitions"));
-    fs.writeFileSync(
-      path.join(projectDir, "workflow_settings.yaml"),
-      `
-defaultProject: ${INTEGRATION_TEST_PROJECT}
-defaultLocation: US
-defaultDataset: df_integration_test_jit_dry_run
-`
-    );
-    fs.writeFileSync(
-      path.join(projectDir, "definitions/jit_table.js"),
-      `publish("jit_table_dry_run", { type: "table" }).jitCode(async (jctx) => "SELECT 1 as id")`
-    );
-
-    // Mock @dataform/core
-    const nodeModulesDir = path.join(projectDir, "node_modules", "@dataform", "core");
-    fs.mkdirpSync(nodeModulesDir);
-    const coreBundlePath = path.resolve("packages/@dataform/core/bundle.js");
-    fs.copyFileSync(coreBundlePath, path.join(nodeModulesDir, "bundle.js"));
-    const corePackageJsonPath = path.resolve("packages/@dataform/core/package.json");
-    fs.copyFileSync(corePackageJsonPath, path.join(nodeModulesDir, "package.json"));
-    fs.writeFileSync(
-      path.join(projectDir, "package.json"),
-      JSON.stringify({ dependencies: { "@dataform/core": "3.0.0-alpha.0" } })
-    );
-
-    try {
-      const compiledGraph = await dfapi.compile({ projectDir });
-      
-      // Drop dataset to start fresh
-      await dbadapter.execute(
-        `drop schema if exists \`${INTEGRATION_TEST_PROJECT}.df_integration_test_jit_dry_run\` cascade`
-      );
-
-      const executionGraph = await dfapi.build(compiledGraph, {}, dbadapter);
-      
-      const runResult = await dfapi.run(dbadapter, executionGraph, { 
-        projectDir, 
-        bigquery: { dryRun: true } 
-      }).result();
-
-      expect(dataform.RunResult.ExecutionStatus[runResult.status]).eql(
-        dataform.RunResult.ExecutionStatus[dataform.RunResult.ExecutionStatus.SUCCESSFUL]
-      );
-
-      // Verify that the table was NOT created
-      const tables = await dbadapter.schemas(INTEGRATION_TEST_PROJECT).then(schemas => {
-        if (!schemas.includes("df_integration_test_jit_dry_run")) {
-          return [];
-        }
-        return dbadapter.tables(INTEGRATION_TEST_PROJECT, "df_integration_test_jit_dry_run");
-      });
-      expect(tables.length).to.equal(0);
-    } finally {
-      if (fs.existsSync(projectDir)) {
-        fs.removeSync(projectDir);
-      }
-    }
-  });
 });
-
-async function cleanWarehouse(
-  compiledGraph: dataform.CompiledGraph,
-  dbadapter: dbadapters.IDbAdapter
-) {
-  await dropAllTables(
-    (await dfapi.build(compiledGraph, {}, dbadapter)).warehouseState.tables,
-    new ExecutionSql(compiledGraph.projectConfig, compiledGraph.dataformCoreVersion),
-    dbadapter
-  );
-}
